feedback fixed

+++ app/controllers/api/v1/authors_controller.rb
@@ -3,10 +3,11 @@ module Api
         class AuthorsController < ApplicationController
             def index
                 authors = Author.paginate(:page => params[:page], :per_page => params[:per_page]).order('created_at desc')
-                render json: {status: 'SUCCESS', message: 'Load authors', data: authors}, status: :ok
+                pagination = { page: params[:page] || 1 , per_page: params[:per_page] || 20, total_pages: authors.total_pages, total_count: authors.total_entries }
+                render json: {status: 'SUCCESS', message: 'Load authors', data: authors, pagination: pagination}, status: :ok
             end
             def show
-                author = Author.find(params[:id])
+                author = Author.find_by(id: params[:id])
                 render json: {status: 'SUCCESS', message: 'Show author', data: author}, status: :ok
             end
             def create 
@@ -18,7 +19,11 @@ module Api
                 end
             end
             def destroy
-                author =  Author.find(params[:id])
+                author = Author.find_by(id: params[:id])
+                if !author
+                    render json: {status: 'FAIL', message: 'Delete author'}, status: :unprocessable_entity
+                    return
+                end
                 if author.destroy
                     render json: {status: 'SUCCESS', message: 'Delete author', data: author}, status: :ok
                 else 
@@ -27,7 +32,11 @@ module Api
             end
 
             def update
-                author = Author.find(params[:id])
+                author = Author.find_by(id: params[:id])
+                if !author 
+                    render json: {status: 'FAIL', message: 'Update author'}, status: :unprocessable_entity
+                    return
+                end
                 if author.update(author_params)
                     render json: {status: 'SUCCESS', message: 'Update author', data: author}, status: :ok
                 else
