+++ app/controllers/api/v1/publishers_controller.rb
@@ -2,12 +2,12 @@ module Api
     module V1 
         class PublishersController < ApplicationController
             def index
-                puplishers = Publisher.paginate(:page => params[:page], :per_page => params[:per_page]).order('created_at desc')
-                render json: {status: 'SUCCESS', message: 'Load puplishers', data: puplishers}, status: :ok
+                publishers = Publisher.paginate(:page => params[:page], :per_page => params[:per_page]).order('created_at desc')
+                pagination = { page: params[:page] || 1 , per_page: params[:per_page] || 20, total_pages: publishers.total_pages, total_count: publishers.total_entries }
+                render json: {status: 'SUCCESS', message: 'Load puplishers', data: publishers, pagination: pagination}, status: :ok
             end
             def show
-                publisher = Publisher.find(params[:id])
-                # books = publisher.books
+                publisher = Publisher.find_by(id: params[:id])
                 render json: {status: 'SUCCESS', message: 'Show book by publisher', data: publisher}, status: :ok
             end
 
@@ -21,16 +21,24 @@ module Api
             end
 
             def destroy
-                publisher = Publisher.find(params[:id])
+                publisher = Publisher.find_by(id: params[:id])
+                if !publisher
+                    render json: {status: 'FAIL', message: 'Delete publisher'}, status: :unprocessable_entity
+                    return
+                end
                 if publisher.destroy
                     render json: {status: 'SUCCESS', message: 'Delete publisher', data: publisher}, status: :ok
                 else 
-                    render json: {status: 'FAIL', message: 'Delete publisher'}, status: :unprocessable_entity
+                    render json: {status: 'FAIL', message: 'Delete publisher', data:publisher.errors}, status: :unprocessable_entity
                 end
             end
 
             def update
-                publisher = Publisher.find(params[:id])
+                publisher = Publisher.find_by(id: params[:id])
+                if !publisher
+                    render json: {status: 'FAIL', message: 'Update publisher'}, status: :unprocessable_entity
+                    return
+                end                
                 if publisher.update(publisher_params)
                     render json: {status: 'SUCCESS', message: 'Update publisher', data: publisher}, status: :ok
                 else
