+++ app/controllers/api/v1/users_controller.rb
@@ -3,16 +3,17 @@ module Api
         class UsersController < ApplicationController
             def index
                 if user_params = ''
-                    users = User.paginate(:page => params[:page], :per_page => params[:per_page]).order('created_at desc')
+                    users = User.paginate(:page => params[:page], :per_page => params[:per_page] || 20).order('created_at desc')
+                    pagination = { page: params[:page] || 1 , per_page: params[:per_page] || 20, total_pages: users.total_pages, total_count: users.total_entries }
                 else
                     users = User.where(role: user_params[:role]).paginate(:page => params[:page], :per_page => params[:per_page]).order('created_at desc')
-                # users = User.paginate(:page => params[:page], :per_page => params[:per_page]).order('created_at desc')
+                    pagination = { page: params[:page] || 1 , per_page: params[:per_page] || 20, total_pages: users.total_pages, total_count: users.total_entries }
                 end
-                render json: {status: 'SUCCESS', message: 'Load user', data: users}, status: :ok
+                render json: {status: 'SUCCESS', message: 'Load user', data: users, pagination: pagination }, status: :ok
             end
 
             def show
-                user = User.find(params[:id])
+                user = User.find_by(id: params[:id])
                 render json: {status: 'SUCCESS', message: 'Show user', data: user}, status: :ok
             end
 
@@ -26,16 +27,24 @@ module Api
             end
 
             def destroy
-                user =  User.find(params[:id])
+                user = User.find_by(id: params[:id])
+                if !user
+                    render json: {status: 'FAIL', message: 'Delete user'}, status: :unprocessable_entity
+                    return
+                end
                 if user.destroy
                     render json: {status: 'SUCCESS', message: 'Delete user', data: user}, status: :ok
                 else 
-                    render json: {status: 'FAIL', message: 'Delete user'}, status: :unprocessable_entity
+                    render json: {status: 'FAIL', message: 'Delete user', data: user.errors}, status: :unprocessable_entity
                 end
             end
 
             def update
-                user = User.find(params[:id])
+                user = User.find_by(id: params[:id])
+                if !user
+                    render json: {status: 'FAIL', message: 'Update user'}, status: :unprocessable_entity
+                    return
+                end                
                 if user.update(user_params)
                     render json: {status: 'SUCCESS', message: 'Update user', data: user}, status: :ok
                 else
